requirements.txt
asgiref==3.8.1
Django==5.1.2
mysqlclient==2.2.5
sqlparse==0.5.1
tzdata==2024.2
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.1.2
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.6
networkx==3.3
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.1
sympy==1.13.0
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.3
